requirements.txt
alembic==1.15.2
asyncpg==0.30.0
fastapi==0.115.12
psycopg2-binary==2.9.10
pydantic-settings==2.8.1
sqlmodel==0.0.24
uvicorn==0.34.0
email-validator==2.2.0
